+++ pkg/cluster/internal/create/actions/createworker/provider.go
@@ -594,7 +594,7 @@ func (p *Provider) deployClusterOperator(n nodes.Node, privateParams PrivatePara
 			c += " --set app.containers.controllerManager.image.tag=" + clusterOperatorImage
 		}
 		if privateParams.Private {
-			c += " --set app.containers.kubeRbacProxy.image=" + keosRegistry.url + "/kube-rbac-proxy:v0.15.0"
+			c += " --set app.containers.kubeRbacProxy.image=" + keosRegistry.url + "/kubebuilder/kube-rbac-proxy:v0.15.0"
 		}
 		if keosCluster.Spec.InfraProvider == "azure" {
 			c += " --set secrets.azure.clientIDBase64=" + strings.Split(p.capxEnvVars[1], "AZURE_CLIENT_ID_B64=")[1] +