+++ CHANGELOG.md
@@ -2,6 +2,9 @@
 
 ## 0.17.0-0.6.2 (2025-02-10)
 
+* [PLT-1628] Fix coredns image registry and repository references
+* [PLT-1628] Fix cluster-api-gcp image registry and repository references
+* [PLT-1628] Fix kube-rbac-proxy image registry and repository references
 * [Core] Fix upgrade process
 
 ## 0.17.0-0.6.1 (2025-01-28)