+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 
 ## 0.17.0-0.7.0 (upcoming)
 
+* [PLT-1394] Bump Flux chart version to 2.14.1
 * [PLT-1393] Bump Tigera Operator version to v3.29.1
 
 ## Previous development